Add CircleCI config
Note that we've disabled MyPy here, this is because mypy is currently failing due to this project being untyped. I want to set CircleCI up to run after merging this PR and don't want it to fail on master. I plan to add typing and re-enable MyPy in a subsequent PR